all other ids are long, but we had a collection of collaboratorIds as List<int> and we were getting errros from zendesk pointing to this. Changed int to long and it is working nicely now.

+++ src/ZendeskApi.Contracts/Models/Request.cs
@@ -64,7 +64,7 @@ namespace ZendeskApi.Contracts.Models
         public long? OrganisationId { get; set; }
 
         [DataMember(Name = "collaborator_ids", EmitDefaultValue = false)]
-        public List<int> CollaboratorIds { get; set; }
+        public List<long> CollaboratorIds { get; set; }
 
         [DataMember(Name = "group_id", EmitDefaultValue = false)]
         public int? GroupId { get; set; }
